+++ CHANGELOG.md
@@ -5,3 +5,4 @@
     - Initialized SortViz project
     - `BaseSorter`: a base class for all the sorters.
     - `BubbleSorter`: a sorter class that will perform the `Bubble Sort` method.
+    - `MergeSorter`: a sorter class that will perform the `Merge Sort` method.